Issue #76 - Better break-up fraction
Almost completely covers all cases of (ax^1 + b) / (cx^1 + d). The only
current holes in this function are for cases where 1) Constant and
polynomial term are not ordered 2) When b is equal to 0

 'checks/canFindDenonimatorInNumerator': The checks are more specific than
before.

'breakUpNumeratorSearch/index.js': Included a multiplier variable to see
how much times the denominator is in the numerator. Instead of using
numerator.args[1] it's now the args[num_n-1] to get the last index, this
way as long as it's sorted we'll always get the constant.

'checks/checks.test.js': More test cases, included falses

'breakUpNumeratorSearch/breakUpNumeratorSearch.test.js': More test cases